backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
together==0.2.7
langchain-community==0.0.38
langchain==0.1.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
tiktoken==0.5.2
python-dotenv==1.0.0
python-multipart==0.0.6